Extracted utility out of definition
