+++ pipeline/src/main/java/com/autobuild/pipeline/utility/file/PipelineFileService.java
@@ -1,4 +1,4 @@
-package com.autobuild.pipeline.definiton.utility.file;
+package com.autobuild.pipeline.utility.file;
 
 import java.io.IOException;
 import java.util.Map;
