+++ pipeline/src/test/java/com/autobuild/pipeline/definition/service/PipelineServiceTest.java
@@ -31,8 +31,8 @@ import com.autobuild.pipeline.definiton.exceptions.DuplicateEntryException;
 import com.autobuild.pipeline.definiton.exceptions.InvalidIdException;
 import com.autobuild.pipeline.definiton.repository.PipelineRepository;
 import com.autobuild.pipeline.definiton.service.PipelineService;
-import com.autobuild.pipeline.definiton.utility.file.PipelineFileService;
 import com.autobuild.pipeline.testutility.DummyData;
+import com.autobuild.pipeline.utility.file.PipelineFileService;
 
 import jakarta.persistence.EntityNotFoundException;
