+++ pipeline/src/test/java/com/autobuild/pipeline/definition/utility/file/extension/ExtensionsTest.java
@@ -5,7 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import org.junit.jupiter.api.Test;
 
-import com.autobuild.pipeline.definiton.utility.file.extension.Extensions;
+import com.autobuild.pipeline.utility.file.extension.Extensions;
 
 public class ExtensionsTest {
     // private Extensions extensions;
